Update to change map, but on logical part

+++ tools/map-visualiser-qt/map-visualiser-qt.h
@@ -100,7 +100,7 @@ private:
     Tiled::Tileset * playerTileset;
     int moveStep;
     Pokecraft::Direction direction;
-    float xPerso,yPerso;
+    quint8 xPerso,yPerso;
     bool inMove;
 
     QTimer timer;
